Option to disable SSL Verification
Currently, I cannot use this the InfluxDB plugin because data is sent without valid SSL certification and our InfluxDB server requires OpenSSL. Adding this simple check to see if the configured host starts with https and then turning off SSL Cert Validation allows me to use the plugin. I have compiled, tested and confirmed that this fix works. A long term check may be to add an SSL cert option to the config file in conjunction with this option.